+++ src/Net/TCNetSocketImp.h
@@ -64,11 +64,11 @@ namespace TC
             virtual bool Close();
             virtual SocketId GetSocket();
 
-            virtual uint32 ReadBytes(void* buffer, uint32 size);
-            virtual uint32 ReadBytes(void* buffer, uint32 size, const Time& timeout);
-            virtual uint32 ReadBytesFrom(void* buffer, uint32 size, Address& ip);
-            virtual uint32 WriteBytes(const void* buffer, uint32 size);
-            virtual uint32 WriteBytesTo(const void* buffer, uint32 size, const Address& ip, PortNumber port);
+            virtual uint64 ReadBytes(void* buffer, uint64 size);
+            virtual uint64 ReadBytes(void* buffer, uint64 size, const Time& timeout);
+            virtual uint64 ReadBytesFrom(void* buffer, uint64 size, Address& ip);
+            virtual uint64 WriteBytes(const void* buffer, uint64 size);
+            virtual uint64 WriteBytesTo(const void* buffer, uint64 size, const Address& ip, PortNumber port);
 
          private:
             SocketId m_socket;
